fix1: 그룹 랭킹 내 인증률 포맷 변경, 그룹 참가 시 API 응답 데이터 추가

+++ src/main/java/site/dogether/challengegroup/controller/response/GetJoiningChallengeGroupTeamActivitySummaryResponse.java
@@ -9,7 +9,7 @@ public record GetJoiningChallengeGroupTeamActivitySummaryResponse(
     public record RankResponse(
         int rank,
         String name,
-        double certificationRate
+        int certificationRate
     ) {
         public static List<RankResponse> of(List<Rank> ranks) {
             return ranks.stream()
